% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{print.invasimapr_fit}
\alias{print.invasimapr_fit}
\title{Compact print}
\usage{
\method{print}{invasimapr_fit}(x, ...)
}
\description{
Compact print
}
\keyword{#}
\keyword{<--}
\keyword{from}
\keyword{hide}
\keyword{index}
\keyword{internal}
