% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_globals.R
\docType{package}
\name{invasimapr-package}
\alias{invasimapr-package}
\title{invasimapr: Workflow to Visualise Trait Dispersion and Assess Invasibility}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilities for assembling invader-resident interaction tensors, computing trait-based competition kernels from distance matrices, deriving environmental suitability kernels, calculating invasion fitness, and summarising trait dispersion metrics. Includes helpers to predict site-by-species responses for residents and simulated invaders, and to derive species invasiveness, site invasibility and trait-effect indicators. The package implements a network invasibility framework that decomposes invasion fitness into abiotic suitability, niche crowding and resident competition components, following the ecological-network theory of Hui and colleagues.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/invasimapr}
  \item \url{https://b-cubed-eu.github.io/invasimapr/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/invasimapr/issues}
}

}
\author{
\strong{Maintainer}: Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})

Authors:
\itemize{
  \item Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})
}

Other contributors:
\itemize{
  \item Stellenbosch University [copyright holder]
}

}
\keyword{internal}
