% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{invasimapr-package}
\alias{invasimapr}
\alias{invasimapr-package}
\title{invasimapr: Workflow to visualise trait dispersion and assess invasibility}
\description{
Utilities for assembling invader–resident interaction tensors, computing trait-based competition kernels from distance matrices, deriving environmental suitability kernels, calculating invasion fitness, and summarising trait dispersion metrics. Includes helpers to predict site × species responses for residents and simulated invaders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/invasimapr}
  \item \url{https://b-cubed-eu.github.io/invasimapr/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/invasimapr/issues}
}

}
\author{
\strong{Maintainer}: Sandra MacFadyen \email{macfadyen@sun.ac.za}

}
\keyword{internal}
