% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_sensitivities.R
\name{learn_sensitivities}
\alias{learn_sensitivities}
\title{Learn sensitivities (αᵢ, βᵢ, θᵢ/γᵢ) and optional site-varying αᵢₛ, Γᵢₛ}
\usage{
learn_sensitivities(fit, use_site_random_slopes = TRUE, lrt = TRUE)
}
\description{
Learn sensitivities (αᵢ, βᵢ, θᵢ/γᵢ) and optional site-varying αᵢₛ, Γᵢₛ
}
