% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Prepare inputs (assemble & align core tables)}
\usage{
prepare_inputs(..., make_plots = FALSE)
}
\value{
\code{invasimapr_fit}
}
\description{
Thin wrapper over \code{assemble_matrices()} that
stores results in an \code{invasimapr_fit} object.
}
