% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise invasiveness/invasibility}
\usage{
summarise_results(
  fit,
  boundary_sf = NULL,
  boundary_params = list(inherit.aes = FALSE, fill = NA, color = "black", size = 0.3),
  traits_inv = NULL,
  ...
)
}
\arguments{
\item{fit}{object with \code{$fitness} and/or \code{$prob}.}

\item{boundary_sf}{An \strong{sf} object to overlay as a boundary on the site map.}

\item{boundary_params}{Named list of aesthetics for \code{ggplot2::geom_sf};
defaults to \code{list(inherit.aes=FALSE, fill=NA, color="black", size=0.3)}.}

\item{...}{forwarded to \code{summarise_invasiveness_invasibility()}.}
}
\value{
updated \code{invasimapr_fit}
}
\description{
Thin wrapper over \code{summarise_invasiveness_invasibility()}.
}
